code interpreter support
