+++ packages/cdk/lib/construct/index.ts
@@ -6,6 +6,5 @@ export * from './rag';
 export * from './transcribe';
 export * from './common-web-acl';
 export * from './agent';
-export * from './file';
 export * from './recognize-file';
 export * from './rag-knowledge-base';
