+++ packages/cdk/lib/generative-ai-use-cases-stack.ts
@@ -9,7 +9,6 @@ import {
   RagKnowledgeBase,
   Transcribe,
   CommonWebAcl,
-  File,
   RecognizeFile,
 } from './construct';
 import { CfnWebACLAssociation } from 'aws-cdk-lib/aws-wafv2';
@@ -155,11 +154,6 @@ export class GenerativeAiUseCasesStack extends Stack {
       hostedZoneId: props.hostedZoneId,
     });
 
-    const file = new File(this, 'File', {
-      userPool: auth.userPool,
-      api: api.api,
-    });
-
     if (ragEnabled) {
       const rag = new Rag(this, 'Rag', {
         userPool: auth.userPool,
@@ -170,7 +164,7 @@ export class GenerativeAiUseCasesStack extends Stack {
       // 既存の Kendra を import している場合、data source が S3 ではない可能性がある
       // その際は rag.dataSourceBucketName が undefined になって権限は付与されない
       if (rag.dataSourceBucketName) {
-        file.allowDownloadFile(rag.dataSourceBucketName);
+        api.allowDownloadFile(rag.dataSourceBucketName);
       }
     }
 
@@ -183,7 +177,7 @@ export class GenerativeAiUseCasesStack extends Stack {
       });
 
       // File API から data source の Bucket のファイルをダウンロードできるようにする
-      file.allowDownloadFile(props.knowledgeBaseDataSourceBucketName!);
+      api.allowDownloadFile(props.knowledgeBaseDataSourceBucketName!);
     }
 
     new Transcribe(this, 'Transcribe', {
@@ -196,7 +190,7 @@ export class GenerativeAiUseCasesStack extends Stack {
       new RecognizeFile(this, 'RecognizeFile', {
         userPool: auth.userPool,
         api: api.api,
-        fileBucket: file.fileBucket,
+        fileBucket: api.fileBucket,
         vpcId: props.vpcId,
       });
     }
