+++ packages/cdk/package.json
@@ -21,16 +21,16 @@
   "dependencies": {
     "@aws-cdk/aws-cognito-identitypool-alpha": "^2.22.0-alpha.0",
     "@aws-cdk/aws-lambda-python-alpha": "2.147.1-alpha.0",
-    "@aws-sdk/client-bedrock-agent-runtime": "^3.549.0",
-    "@aws-sdk/client-bedrock-runtime": "^3.549.0",
-    "@aws-sdk/client-dynamodb": "^3.549.0",
-    "@aws-sdk/client-kendra": "^3.549.0",
-    "@aws-sdk/client-s3": "^3.549.0",
-    "@aws-sdk/client-sagemaker-runtime": "^3.549.0",
-    "@aws-sdk/client-transcribe": "^3.549.0",
-    "@aws-sdk/client-transcribe-streaming": "^3.549.0",
-    "@aws-sdk/lib-dynamodb": "^3.549.0",
-    "@aws-sdk/s3-request-presigner": "^3.549.0",
+    "@aws-sdk/client-bedrock-agent-runtime": "^3.614.0",
+    "@aws-sdk/client-bedrock-runtime": "^3.614.0",
+    "@aws-sdk/client-dynamodb": "^3.614.0",
+    "@aws-sdk/client-kendra": "^3.614.0",
+    "@aws-sdk/client-s3": "^3.614.0",
+    "@aws-sdk/client-sagemaker-runtime": "^3.614.0",
+    "@aws-sdk/client-transcribe": "^3.614.0",
+    "@aws-sdk/client-transcribe-streaming": "^3.614.0",
+    "@aws-sdk/lib-dynamodb": "^3.614.0",
+    "@aws-sdk/s3-request-presigner": "^3.614.0",
     "@aws-solutions-constructs/aws-cloudfront-s3": "^2.54.1",
     "aws-cdk-lib": "2.147.1",
     "constructs": "^10.3.0",
