+++ packages/web/src/components/Markdown.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import { BaseProps } from '../@types/common';
 import { default as ReactMarkdown } from 'react-markdown';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
@@ -8,6 +8,7 @@ import remarkBreaks from 'remark-breaks';
 import ButtonCopy from './ButtonCopy';
 import useRagFile from '../hooks/useRagFile';
 import { PiSpinnerGap } from 'react-icons/pi';
+import useFileApi from '../hooks/useFileApi';
 
 type Props = BaseProps & {
   children: string;
@@ -50,6 +51,20 @@ const LinkRenderer = (props: any) => {
     </>
   );
 };
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const ImageRenderer = (props: any) => {
+  const { isS3Url } = useRagFile();
+  const { getDocDownloadSignedUrl } = useFileApi();
+  const [src, setSrc] = useState(props.src);
+
+  useEffect(() => {
+    if (isS3Url(props.src)) {
+      getDocDownloadSignedUrl(props.src).then((url) => setSrc(url));
+    }
+  }, [getDocDownloadSignedUrl, isS3Url, props.src]);
+
+  return <img id={props.id} src={src} />;
+};
 
 const Markdown: React.FC<Props> = ({ className, prefix, children }) => {
   return (
@@ -62,6 +77,7 @@ const Markdown: React.FC<Props> = ({ className, prefix, children }) => {
       remarkRehypeOptions={{ clobberPrefix: prefix }}
       components={{
         a: LinkRenderer,
+        img: ImageRenderer,
         sup: ({ children }) => (
           <sup className="m-0.5 rounded-full bg-gray-200 px-1">{children}</sup>
         ),
